refactor: move mood diary client into the new api folder
